kernel cmdline: enable deb package patching cmdline
airtop-fpsvc daemon requires i2c-i801 kernel module, which in turn
requires "acpi_enforce_resources=lax" being present in the kernel
command line.
This commit checks /etc/default/grub file and adds the required suffix
in case it is not there already.
Note:
We do not remove the suffix upon package removal, since it is required
anyhow by the i2c-i801 driver on AirTop.
The reason i2c-i801 is not loaded automatically upon boot is not
clear, and might as well turn out to be a bug.

Signed-off-by: Andrey Gelman <andrey.gelman@compulab.co.il>